refactor(framework): daemon.ts is the lifecycle, daemon-runtime.ts is the work
createProjectRuntime was a 360-line closure inside daemon.ts -- longer than
the daemon-services module #935 extracted beside it, and its own doc
claimed a split ('so the daemon body reads as lifecycle and this reads as
business logic') that the file layout never finished. daemon.ts drops from
948 lines to 417: the state file, ports, boot and shutdown. The runtime,
and the spawn/terminate plumbing only it and ensureDaemon share, move to
daemon-runtime.ts as a pure move.

Two consolidations ride the move because the move is what exposed them:

- The '<projectKey>::<runId>' composite key was hand-built in onStart,
  hand-prefix-matched in activeRunCount, hand-split in suspendRuns, and
  re-implemented as previewKeyFor. One scopedKey/parseScopedKey/
  keyBelongsTo trio now owns the encoding.
- daemon.ts's isProcessAlive was byte-identical to the store's isPidAlive,
  both exported. One body remains in the store; the daemon re-exports it
  under its historical public name, so the npm surface is unchanged.

startOptionFlags is re-exported from daemon.ts likewise; no import site
outside the package changes.
